server/pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = ""
authors = [{name = "Igor Borges", email = "igor@borges.dev"}]
readme = "README.md"

requires-python = ">=3.10"

dependencies = [
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0", # No change
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-experimental>=0.3.4",
    "langchain-google-community>=2.0.7",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "langserve>=0.3.1",
    "matrix-nio[e2e]>=0.25.2", # No change
    "neo4j>=5.28.1",
    "pygithub>=2.6.1",
    "sse-starlette>=2.3.5",
    "uvicorn>=0.34.2",
    "wikipedia>=1.4.0", # No change
    "zep-cloud>=2.12.3",
]
